FIX: added whispey landing page and sign up page fixes WIP
